chore: move get input override to a function
