+++ src/commands/call.ts
@@ -1,14 +1,12 @@
-import { readFile } from 'node:fs/promises';
-import { resolve } from 'node:path';
 import process from 'node:process';
 
 import { Args, Flags } from '@oclif/core';
 import { ActorStartOptions, ApifyClient } from 'apify-client';
 
 import { ApifyCommand } from '../lib/apify_command.js';
+import { getInputOverride } from '../lib/commands/resolve-input.js';
 import { SharedRunOnCloudFlags, runActorOrTaskOnCloud } from '../lib/commands/run-on-cloud.js';
-import { CommandExitCodes, LOCAL_CONFIG_PATH } from '../lib/consts.js';
-import { error } from '../lib/outputs.js';
+import { LOCAL_CONFIG_PATH } from '../lib/consts.js';
 import { getLocalConfig, getLocalUserInfo, getLoggedClientOrThrow } from '../lib/utils.js';
 
 export class ActorCallCommand extends ApifyCommand<typeof ActorCallCommand> {
@@ -77,7 +75,7 @@ export class ActorCallCommand extends ApifyCommand<typeof ActorCallCommand> {
             runOpts.memory = this.flags.memory;
         }
 
-        const inputOverride = await this.resolveInputOverride(cwd);
+        const inputOverride = await getInputOverride(cwd, this.flags.input, this.flags.inputFile);
 
         // Means we couldn't resolve input, so we should exit
         if (inputOverride === false) {
@@ -92,7 +90,7 @@ export class ActorCallCommand extends ApifyCommand<typeof ActorCallCommand> {
             runOptions: runOpts,
             type: 'Actor',
             waitForFinishMillis,
-            inputOverride,
+            inputOverride: inputOverride?.input,
         });
     }
 
@@ -144,85 +142,5 @@ export class ActorCallCommand extends ApifyCommand<typeof ActorCallCommand> {
         }
 
         throw new Error('Please provide an Actor ID or name, or run this command from a directory with a valid Apify Actor.');
-    }
-
-    private async resolveInputOverride(cwd: string) {
-        let input: Record<string, unknown> | undefined;
-
-        if (!this.flags.input && !this.flags.inputFile) {
-            // Try reading stdin
-            const stdin = await this.readStdin(process.stdin);
-
-            if (stdin) {
-                try {
-                    const parsed = JSON.parse(stdin);
-
-                    if (Array.isArray(parsed)) {
-                        throw new Error('The input provided is invalid. It should be an object, not an array.');
-                    }
-
-                    input = parsed;
-                } catch (err) {
-                    error({ message: `Cannot parse JSON input from standard input.\n  ${(err as Error).message}` });
-                    process.exitCode = CommandExitCodes.InvalidInput;
-                    return false;
-                }
-            }
-
-            return input;
-        }
-
-        if (this.flags.input) {
-            switch (this.flags.input[0]) {
-                case '-': {
-                    error({ message: 'You need to pipe something into standard input when you specify the `-` value to `--input`.' });
-                    process.exitCode = CommandExitCodes.InvalidInput;
-                    return false;
-                }
-                default: {
-                    try {
-                        const parsed = JSON.parse(this.flags.input);
-
-                        if (Array.isArray(parsed)) {
-                            throw new Error('The input provided is invalid. It should be an object, not an array.');
-                        }
-
-                        input = parsed;
-                    } catch (err) {
-                        error({ message: `Cannot parse JSON input.\n  ${(err as Error).message}` });
-                        process.exitCode = CommandExitCodes.InvalidInput;
-                        return false;
-                    }
-                }
-            }
-        } else if (this.flags.inputFile) {
-            switch (this.flags.inputFile[0]) {
-                case '-': {
-                    error({ message: 'You need to pipe something into standard input when you specify the `-` value to `--input-file`.' });
-                    process.exitCode = CommandExitCodes.InvalidInput;
-                    return false;
-                }
-                default: {
-                    const fullPath = resolve(cwd, this.flags.inputFile);
-
-                    try {
-                        const fileContent = await readFile(fullPath, 'utf8');
-                        const parsed = JSON.parse(fileContent);
-
-                        if (Array.isArray(parsed)) {
-                            throw new Error('The input provided is invalid. It should be an object, not an array.');
-                        }
-
-                        input = parsed;
-                    } catch (err) {
-                        error({ message: `Cannot read input file at path "${fullPath}".\n  ${(err as Error).message}` });
-                        process.exitCode = CommandExitCodes.InvalidInput;
-                        return false;
-                    }
-                }
-            }
-        }
-
-        return input;
     }
 }
